src/main/java/com/john/Nov2022/leetcode/q700_799/Question704.java
package com.john.Nov2022.leetcode.q700_799;

public class Question704 {
    public int search(int[] nums, int target) {
        if (nums == null || nums.length == 0) {
            return -1;
        }

        return binarySearch(nums, target, 0, nums.length - 1);
    }

    private int binarySearch(int[] nums, int target, int left, int right) {
        if (right - left <= 1) {
            if (nums[right] == target) {
                return right;
            } else if (nums[left] == target) {
                return left;
            } else {
                return -1;
            }
        }

        int mid = left + (right - left) / 2;

        if (nums[mid] == target) {
            return mid;
        } else if (nums[mid] > target) {
            return binarySearch(nums, target, left, mid - 1);
        } else {
            return binarySearch(nums, target, mid + 1, right);
        }
    }
}
